Add evaluation of if expressions, test factorial func!
Note that this required moving the function application production to be
above all the arithmetic. Otherwise with an input like:

factorial = \ n { if n == 0 then 1 else n * factorial(n - 1) }

the parser would bind the 'n * factorial' first, resulting in an error.
The 'factorial(n - 1)' needed to be bound and then have the
multiplication production after that.