Adjust grammar for idfplus object definitions, add member access
Previously, the name for a property of an object was a 'STRING' token.
This generally made the grammar match that of JSON.

But we aren't JSON, and I don't think it makes sense to deal with all
the issues that would arise with how to handle a string as the
identifier.

This commit also added the most basic required logic for member access
on objects. It adds the first example of mapping over a list of objects,
a very common expected usage of the language.