Add 'load' function
This is a critical function - it will allow for extremely easy loading
of data from other files. The first implementation is just reading from
a tab delimited file, with a relative path passed as a StringExpression
argument.

In the future, however, this is going to be massively expanded, so that
it will be able to read directly from spreadsheet or other file types.
EPPlus is going to be brought in sometime soon to read directly from
Excel.