Update export grammar, IDENTIFIER token
So using '.' as the separator for the namespace name was a terrible
idea. It creates instant ambiguity between whether you wanted a
namespaced variable, or a variable with a member access.

I decided to use '@' as the namespace separator because I think it looks
cool, but mostly because it was one of the few characters that hadn't
been used up in the grammar yet.

Also wrap up all the identifier for the export rule, just for
consistency sake with the 'only' and 'not' rules. I actually don't think
the parenthesis are required at all for any of them, but it is nicer to
look at them grouped?