Add range and sheet options for loading data from Excel
Used ANTLR for parsing the Excel range syntax. Currently offer 3 forms:

 1. 'A1:B2' - Complete Range
 2. 'A1'    - Just the start cell
 3. '1:A:C' - Start row with columns specified (like excelchop)

Only 1 and 2 are implemented.

This also required being more explicit about the working directory
while evaluating expressions, since the path can be relative, and it's
important to know what location the file is in.

This also added an extension method for sanitizing identifiers supplied
by the user. This is used when parsing the headers of data.