Add class for comparing two 'idf' files
It's most important that the idf objects that EnergyPlus would actually
see are the same from my unit tests are the same. The final formatting
and whitespacing of the output is going to probably be in flux for a
long time. What I need to check ignores comments and does case
insensitive field checks, also trimming whitespace.