Add boolean comparisons for booleans, empty lists, and count diff lists
Surprisingly enough, I never had gotten around to implementing boolean
comparisons for booleans. Without that, you couldn't do something like

function(asdf) == false

Which is required for checking for false, since right now the grammar
does not have unary inversion for booleans.

This also adds a simple comparison for empty lists so that you can do
checks like

list == []

instead of

length(list) == 0

The list to list equality check will also short circuit to false if the
length of the lists are not equal. The equality comparisons for lists
that are the same length, with the length greater than one, still has
not been implemented. Before that happens, I'm going to have to add some
equality methods to the base types themselves.