Add abstraction for main formatter Visitor
The visitor for the main file and within functions are very similar.
This was an attempt to consolidate the two. It isn't currently being
used, so this was just to commit it in for historical sake, in case I
want to come back to it, without storing it in a long term branch.