Add map and pipe operator
The map operator was defined as '->' and the pipe operator was defined
as '|>'. I wanted to use a single vertical bar for the pipe like shell
syntax, but that is already used by the inline tables. I don't think it
was going to be worth trying to figure out how to handle the ambiguity,
and '|>' seems like an appropriate compromise.